% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{hm_inform}
\alias{hm_inform}
\title{Emit package messages honoring options(healthmarkers.verbose)}
\usage{
hm_inform(level = c("inform", "debug"), msg)
}
\arguments{
\item{level}{"inform" or "debug"}

\item{msg}{character(1)}
}
\description{
Emit package messages honoring options(healthmarkers.verbose)
}
\keyword{internal}
