% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{hm_validate_inputs}
\alias{hm_validate_inputs}
\title{HM-CS v2 input validation hook}
\usage{
hm_validate_inputs(data, col_map, required_keys, fn)
}
\arguments{
\item{data}{data.frame}

\item{col_map}{named list or NULL}

\item{required_keys}{character vector of required keys}

\item{fn}{function name (string)}
}
\description{
HM-CS v2 input validation hook
}
\keyword{internal}
