% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_risk.R
\name{cvd_risk}
\alias{cvd_risk}
\title{Compute cardiovascular risk by selected model}
\usage{
cvd_risk(
  data,
  model = c("ASCVD", "QRISK3", "MESA", "Stroke", "WHO", "RiskScorescvd"),
  year = 10,
  ...
)
}
\arguments{
\item{data}{A tbl data.frame with required columns.}

\item{model}{One of c("ASCVD","QRISK3","MESA","Stroke","WHO","RiskScorescvd").}

\item{year}{Risk horizon for applicable models (10 or 30).}

\item{...}{Additional args passed to underlying wrapper.}
}
\description{
Dispatch to the appropriate risk function
}
