% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_risk.R
\name{cvd_risk_qrisk3}
\alias{cvd_risk_qrisk3}
\title{QRISK3 10-year risk (UK QRISK3-2017)}
\usage{
cvd_risk_qrisk3(data, ...)
}
\arguments{
\item{data}{A tibble with columns required by QRISK3 (see package documentation).}

\item{...}{Additional arguments for QRISK3_2017.}
}
\value{
A tibble with columns \code{model}, \code{year} (always 10), and \code{risk}.
}
\description{
Wrapper around QRISK3::QRISK3_2017.
}
\examples{
\dontrun{
library(tibble)
df <- tibble(age=55, sex=1, sbp=140, HDL_c=50, total_chol=200, smoker=FALSE, diabetes=FALSE)
cvd_risk_qrisk3(df)
}
}
