% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_risk.R
\name{cvd_risk_scorescvd}
\alias{cvd_risk_scorescvd}
\title{RiskScorescvd calculator}
\usage{
cvd_risk_scorescvd(data, ...)
}
\arguments{
\item{data}{Data required by \code{RiskScorescvd::calc_scores()}.}

\item{...}{Passed to \code{RiskScorescvd::calc_scores()}.}
}
\value{
Object returned by \code{RiskScorescvd::calc_scores()}.
}
\description{
Passthrough to \code{RiskScorescvd::calc_scores()} with graceful fallback to NA.
}
