% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{normalize_vec}
\alias{normalize_vec}
\title{Normalize a numeric vector}
\usage{
normalize_vec(x, method = c("none", "z", "inverse", "range", "robust"))
}
\arguments{
\item{x}{Numeric vector}

\item{method}{One of c("none","z","inverse","range","robust")}
}
\value{
Numeric vector of same length as x
}
\description{
Normalize a numeric vector
}
\keyword{internal}
