% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate that col_map is fully specified and data has the mapped columns}
\usage{
validate_inputs(data, col_map, fun_name = "", required_keys = names(col_map))
}
\arguments{
\item{data}{A data.frame or tibble}

\item{col_map}{Named list of expected keys -> column names}

\item{fun_name}{Character; name of the calling function for messages}

\item{required_keys}{Character vector; names in col_map that are required}
}
\description{
Validate that col_map is fully specified and data has the mapped columns
}
\keyword{internal}
